Add wheel to requirements, try running with pypy
Installed pypy with pyenv, ran it, still 105/51 so no increase. Also 105/51 is _worse_ than 106/50, not sure how I thought this was an improvement over no optimization (-OO).
Running it again it performed worse again, only 103/53. Maybe .1 seconds is already so much time that fast execution doesn't give it a benefit but I kinda doubt that. Maybe I'm also not invoking or measuring things correctly, this was just a random attempt.