fuck. me.
After training for ~2000 epochs, train loss has stagnated pretty well and accuracy improvements seem to be slowing down too. However, this is all on train, since val is also train actually. I may have been overfitting the entire time and not noticing.
Next steps are to reduce model complexity again and try to get it to overfit to a small portion of the dataset.

Max accuracy achieved was .07 so 7% which is miserable...