Compare binary payoff vs points payoff
Binary outperforms point-based by a tiny bit (80/77) with std 36 across 1000 games.
Therefore it'll be the new default.

The rational behind this was that the bot should learn to win with the highest probability, not to achieve the highest number of points. If we are playing multiple games and scores are added up, then prioritizing points may make sense. Interesting note is, that the binary payoff achieved a higher mean score (I didn't actually check the win rate) so it's seemingly better no matter what.