Simplify ISMCTS class to avoid callables in the object
Callables cannot be pickled, meaning a multiprocessing ISMCTS implementation is a lot harder. Since we're near the end of the project and have already done a considerable amount of tuning on the different payoff functions, etc. we can set those into stone as they won't be iterated upon at this stage anymore anyway.